Split sandwich view reducer from container code
